requirements.txt
livekit >= 0.18.0
livekit-agents~=0.11.0
livekit-plugins-openai~=0.10.5
livekit-plugins-deepgram~=0.6.9  # Add this line
livekit-plugins-silero~=0.7.3    # Add this line
python-dotenv~=1.0
pdfplumber~=0.11.4
openai~=1.53.0
python-pptx~=1.0.2
pillow~=10.3.0
pytesseract~=0.3.13
tenacity~=9.0.0
httpx~=0.27.2
google-cloud-texttospeech~=2.21.0
sqlalchemy~=2.0.36
fastapi~=0.115.4
psycopg2-binary~=2.9.10
alembic~=1.14.0